enhance: Translate of commands #34

+++ cogs/pin.py
@@ -147,7 +147,7 @@ offにするには、`n!pin off`と送信してください。
             await ctx.reply("あなたには権限がありません。")
             return
 
-    @nextcord.message_command(name="下部ピン留めする", guild_ids=n_fc.GUILD_IDS)
+    @nextcord.message_command(name="Set BottomPin", name_localizations={nextcord.Locale.ja: "下部ピン留めする"} ,guild_ids=n_fc.GUILD_IDS)
     async def pin_message_command(self, interaction: Interaction, message: nextcord.Message):
         if not admin_check.admin_check(interaction.guild, interaction.user):
             await interaction.response.send_message("あなたには管理者権限がありません。", ephemeral=True)
@@ -161,11 +161,11 @@ offにするには、`n!pin off`と送信してください。
         await database.default_push(self.bot.client, pin_message)
         await interaction.response.send_message(f"ピン留めを保存しました。", embed=nextcord.Embed(title="ピン留め", description=message.content), ephemeral=True)
 
-    @nextcord.slash_command(name="pin", description="メッセージ下部ピン留め機能", guild_ids=n_fc.GUILD_IDS)
+    @nextcord.slash_command(name="pin", description="BottomUp command", guild_ids=n_fc.GUILD_IDS)
     async def pin_slash(self, interaction: Interaction):
         pass
 
-    @pin_slash.subcommand(name="on", description="メッセージ下部ピン留め機能をONにする")
+    @pin_slash.subcommand(name="on", description="Turn ON the bottom pin message", description_localizations={nextcord.Locale.ja: "下部ピン留めをONにする"})
     async def on_slash(
         self,
         interaction: Interaction
@@ -176,7 +176,7 @@ offにするには、`n!pin off`と送信してください。
             await interaction.response.send_message("あなたには管理者権限がありません。", ephemeral=True)
             return
 
-    @pin_slash.subcommand(name="off", description="メッセージ下部ピン留め機能をOFFにする")
+    @pin_slash.subcommand(name="off", description="Turn OFF the bottom pin message", description_localizations={nextcord.Locale.ja: "下部ピン留めをOFFにする"})
     async def off_slash(
         self,
         interaction: Interaction,
@@ -214,22 +214,28 @@ offにするには、`n!pin off`と送信してください。
                     except Exception:
                         continue
                     continue
-                try:
-                    if CHANNEL.last_message is not None:
-                        if (CHANNEL.last_message.content == pin_message.value[i][j] and CHANNEL.last_message.author.id == self.bot.user.id):
-                            continue
-                except Exception:
-                    logging.error(traceback.format_exc())
-                    continue
-                messages = await CHANNEL.history(limit=10).flatten()
-                for message in messages:
-                    if message.content == pin_message.value[i][j] and message.author.id == self.bot.user.id:
-                        try:
-                            await message.delete()
-                        except Exception:
-                            logging.error(traceback.format_exc())
-                            continue
-                await CHANNEL.send(pin_message.value[i][j])
+                if j not in self._queue:
+                    self._queue[j] = CHANNEL.last_message
+
+                if self._queue is not None:
+                    try:
+                        if CHANNEL.last_message is not None:
+                            if (CHANNEL.last_message.content == pin_message.value[i][j] and CHANNEL.last_message.author.id == self.bot.user.id):
+                                continue
+                    except Exception:
+                        logging.error(traceback.format_exc())
+                        continue
+                    messages = await CHANNEL.history(limit=10).flatten()
+                    for message in messages:
+                        if message.content == pin_message.value[i][j] and message.author.id == self.bot.user.id:
+                            try:
+                                await message.delete()
+                            except Exception:
+                                logging.error(traceback.format_exc())
+                                continue
+                    await CHANNEL.send(pin_message.value[i][j])
+                    self._queue[j] = None
+
 
 # logging.error(traceback.format_exc())
